feat(Processor): add still diverged state
The PseudoBody now has an additional event state to know if the
source is still diverged with the character. This is similar to
the Unity collision pattern of Enter, Stay, Exit. So now the
divergence events are:

* Diverged (called once the first time divergence occurs)
* StillDiverged (called every frame it is still diverged)
* Converged (called once when divergence is no longer occuring)

This now enables better checking of divergence state and helps solve
body collisions with continuous movements.