fix(Processor): use offset scale when calculating height
There was an issue when the offset scale was larger than 1 then the
height calculation would cause the rigidbody position to be set
incorrectly which would infinitely increase the y position of the
character controller. This would cause the target to be flung in the
air. This fix ensures the height is calculated from the offset Y
position but taking into consideration any scaling on the offset.