feat(Processor): add event when a divergence will occur
The new WillDiverge event will emit when the new `CheckWillDiverge`
method is called passing in a potential target position to move the
pseudo body to. If the given position will cause a divergence then
the WillDiverge event is emitted.

This can be used to determine if an external movement of the
Facade.Source or Facade.Offset may cause a divergence and therefore
can be prevented to stop any collision clipping from occurring.

The Character radius is now also exposed on the Facade and is set to
a default of `0.3` along with the Source Thickness being defaulted
to `0.3` and the Divergence Threshold being reduced to `0.001`.

The PseudoBodyProcessor has also been updated to have the option to
smooth damp any movement of the Facade.Source or Facade.Offset for
any custom effects that may be required. By default, these damp
durations are both set to 0.

A large chunk of refactoring has been done too, so the methods are
broken up into more manageable and overridable chunks.